Add files via upload
A flash card maker that has three modes with flash cards that the user creates in the beginning, Easy: press enter to skip past the current word and see the definition, Moderate: type the correct word matching up to the definition, Hard: type the definition matching up to the term. Each mode has a couple of tries depending on the word/definition. A hint feature gives the letters one at a time after 2 fails, until you run out of tries or get it right. A scoring system is in place as well and you can choose to continue playing after a game is finished without having to leave.